Casts space to u64 for rent calculation
Ensures rent calculation uses u64 for space, preventing potential overflow issues when calculating rent exemption and minimum balance, by explicitly casting the `space` variable to `u64`.

+++ lang/syn/src/codegen/accounts/constraints.rs
@@ -391,7 +391,7 @@ pub fn generate_constraint_rent_exempt(
     match c {
         ConstraintRentExempt::Skip => quote! {},
         ConstraintRentExempt::Enforce => quote! {
-            if !satellite_lang::arch_program::rent::Rent::is_exempt(#info.lamports(), #info.try_data_len()?) {
+            if !satellite_lang::arch_program::rent::Rent::is_exempt(#info.lamports(), #info.try_data_len()? as u64) {
                 return Err(satellite_lang::error::Error::from(satellite_lang::error::ErrorCode::ConstraintRentExempt).with_account_name(#name_str));
             }
         },
@@ -423,7 +423,7 @@ fn generate_constraint_realloc(
         }
 
         let __field_info = #field.to_account_info();
-        let __new_rent_minimum = satellite_lang::arch_program::rent::Rent::minimum_balance(#new_space);
+        let __new_rent_minimum = satellite_lang::arch_program::rent::Rent::minimum_balance(#new_space as u64);
 
         let __delta_space = (::std::convert::TryInto::<isize>::try_into(#new_space).unwrap())
             .checked_sub(::std::convert::TryInto::try_into(__field_info.data_len()).unwrap())
@@ -1117,7 +1117,7 @@ fn generate_constraint_init_group(
                         }
 
                         {
-                            let required_lamports = satellite_lang::arch_program::rent::Rent::minimum_balance(space);
+                            let required_lamports = satellite_lang::arch_program::rent::Rent::minimum_balance(space as u64);
                             if pa.to_account_info().lamports() < required_lamports {
                                 return Err(satellite_lang::error::Error::from(satellite_lang::error::ErrorCode::ConstraintRentExempt).with_account_name(#name_str));
                             }
@@ -1700,7 +1700,7 @@ fn generate_create_account(
         if __current_lamports == 0 {
             // Create the token account with right amount of lamports and space, and the correct owner.
             let space = #space;
-            let lamports = satellite_lang::arch_program::rent::Rent::minimum_balance(space);
+            let lamports = satellite_lang::arch_program::rent::Rent::minimum_balance(space as u64);
             let cpi_accounts = satellite_lang::system_program::CreateAccount {
                 from: #payer.to_account_info(),
                 to: #field.to_account_info()
@@ -1710,7 +1710,7 @@ fn generate_create_account(
         } else {
             require_keys_neq!(#payer.key(), #field.key(), satellite_lang::error::ErrorCode::TryingToInitPayerAsProgramAccount);
             // Fund the account for rent exemption.
-            let required_lamports = satellite_lang::arch_program::rent::Rent::minimum_balance(#space)
+            let required_lamports = satellite_lang::arch_program::rent::Rent::minimum_balance(#space as u64)
                 .max(1)
                 .saturating_sub(__current_lamports);
             if required_lamports > 0 {
